Moved the conversion and export files, for correct npm build and package implementation.
